# Fixed total performance count in PerformanceMonitor