Performance count per agent was fetched and always overwritten. But the overwritten value was just from the first worker and not summed up (like the total value).

Layout:
```
INFO  17:06:04 Total number of agents: 2
INFO  17:06:04 Total number of Hazelcast member workers: 4
INFO  17:06:04 Total number of Hazelcast client workers: 0
INFO  17:06:04     Agent 172.16.16.206 members: 2 clients: 0
INFO  17:06:04     Agent 172.16.16.209 members: 2 clients: 0
```

Before:
```
INFO  17:07:28 Total performance    100.00%         596,803 ops       19,893.43 ops/s
               Agent 172.16.16.209   24.38%         145,488 ops        4,849.60 ops/s
               Agent 172.16.16.206   25.00%         149,187 ops        4,972.90 ops/s
```

After:
```
INFO  17:00:32 Total performance      100.00%         575,700 ops       19,190.00 ops/s
INFO  17:00:32   Agent 172.16.16.209   49.30%         283,809 ops        9,460.30 ops/s
INFO  17:00:32   Agent 172.16.16.206   50.70%         291,891 ops        9,729.70 ops/s
```